Major overhaul to the way Queries work
-QueryHandler no longer internally tracks subclassing of itself. This behavior worked alright, but wouldn't be compatible going forward.
Instead, the HandlerRegistry class is used.
-Created the HandlerRegistry class, which keeps track of all handlers. Eventually this will be a manager as well, allowing for the
grouping and selection of Handlers to provide to a VI. Currently it requires #load_handlers! to be run, provided with a directory. It then
loads all ruby files in that directory.
-Consequently, VIBase#load_handlers! and related functions have been removed. When a VI is initialized with no arguments, it now simply
loads all the handlers that the HandlerRegistry has loaded. Make sure to load your handlers before creating a VI
-Added the Module level function #create_handler, which takes in an optional Superclass, and a block. Evaluates the block in the context
of a new class (subclassed by default from DismalTony::QueryHandler) that is then registered as a handler in the handler registry. This is
the new way to define a handler, by writing a file containing one or more DismalTony#create_handler directives.